% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMatMult}
\alias{eigenMatMult}
\title{Fast matrix multiplication using RcppEigen}
\usage{
eigenMatMult(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}
}
\value{
The matrix product of A times B
}
\description{
Fast matrix multiplication using RcppEigen
}
\examples{
A = matrix(0, ncol = 500, nrow = 500)
system.time(A \%*\% A)
system.time(eigenMatMult(A, A))
}
