% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenSpMatMult}
\alias{eigenSpMatMult}
\title{Fast sparse matrix multiplication using RcppEigen}
\usage{
eigenSpMatMult(A, B)
}
\arguments{
\item{A}{a dgCMatrix}

\item{B}{a dgCMatrix}
}
\value{
The matrix product of A times B
}
\description{
Fast sparse matrix multiplication using RcppEigen
}
\examples{
library(Matrix)
n = 1000
A = matrix(rpois(n*n, 1), nrow = n)
system.time(A \%*\% A)
A = as(A, "dgCMatrix")
system.time(eigenSpMatMult(A, A))
}
