% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenSpResidop}
\alias{eigenSpResidop}
\title{fast sparse matrix residual operator using RcppEigen}
\usage{
eigenSpResidop(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}
}
\value{
The matrix product of
\deqn{A - B(B^t B)^{-1} B^t A}
}
\description{
fast sparse matrix residual operator using RcppEigen
}
\examples{
Y = M = diag(1, 500)
system.time(ruv::residop(Y, M))
Y = as(Y, "dgCMatrix")
M = as(M, "dgCMatrix")
system.time(scMerge::eigenSpResidop(Y, M))
}
