% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{Subsetted mouse ESC `SingleCellExperiment` object}
\format{
A `SingleCellExperiment` object
}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-2600/}
}
\usage{
data(example_sce, package = 'scMerge')
}
\description{
A dataset containing 300 cells and 2026 genes 
from two batches of mouse ESC data
}
\references{
Kolodziejczyk et al.
}
\keyword{datasets}
