% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvSimulate.R
\name{ruvSimulate}
\alias{ruvSimulate}
\title{Simulate a simple matrix}
\usage{
ruvSimulate(m = 100, n = 1e+05, nc = 1000, nRep = 50,
  lambda = 0.1)
}
\arguments{
\item{m}{Number of observations}

\item{n}{Number of features}

\item{nc}{Number of negative controls}

\item{nRep}{Number of replicates}
}
\description{
Simulate a simple matrix
}
\examples{
L = ruvSimulate(m = 700, n = 10000, nRep = 10, lambda = 0.1)
Y = L$Y; M = L$M; ctl = L$ctl; batch = L$batch

microbenchmark::microbenchmark(
improved1 = scMerge::fastRUVIII(Y = Y, M = M, ctl = ctl, k = 20, fast_svd = FALSE),
old = ruv::RUVIII(Y = Y, M = M, ctl = ctl, k = 20),
times = 1)

microbenchmark::microbenchmark(scMerge::scRUVIII(Y = Y, M = M, ctl = ctl, k = 20, batch = batch), times = 1)
microbenchmark::microbenchmark(fakePackage::scRUVIII(Y = Y, M = M, ctl = ctl, k = 20, batch = batch), times = 1)

}
