% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{sce_cbind}
\alias{sce_cbind}
\title{Combine SingleCellExperiment Object.}
\usage{
sce_cbind(sce_list, method = "union", cut_off_batch = 0.01,
  cut_off_overall = 0.01, exprs = c("counts", "logcounts"),
  colData_names = NULL, batch_names = NULL)
}
\arguments{
\item{sce_list}{A list contains the \code{SingleCellExperiment} Object from each batch}

\item{method}{A string indicates the method of combining the gene expression matrix,
either \code{union} or \code{intersect}}

\item{cut_off_batch}{A numeric indicates cut-off for the proportion of a gene is expressed within each batch}

\item{cut_off_overall}{A numeric indicates the cut-off for the proportion of a gene is expressed overall}

\item{exprs}{A string vector indicates the exprssion matrix that are combined}

\item{colData_names}{A string vector indicates the \code{colData} that are combined}

\item{batch_names}{A string vector indicates the batch names}
}
\value{
A combined \code{SingleCellExperiment} objects
}
\description{
Combind the \code{SingleCellExperiment} objects from different batches/experiments
}
\examples{
require(SingleCellExperiment)
data("sce_mESC")
batchNames<-unique(sce_mESC$batch)
sce_list<-list(sce_mESC[,sce_mESC$batch==batchNames[1]],
               sce_mESC[,sce_mESC$batch==batchNames[2]],
               sce_mESC[,sce_mESC$batch==batchNames[3]],
               sce_mESC[,sce_mESC$batch==batchNames[4]],
               sce_mESC[,sce_mESC$batch==batchNames[5]])
sce_combine<-sce_cbind(sce_list,batch_names=batchNames)
}
\author{
Yingxin Lin
}
