% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_mESC}
\alias{sce_mESC}
\title{mouse ESC `SingleCellExperiment` object}
\format{A `SingleCellExperiment` object}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-2600/}
}
\usage{
sce_mESC
}
\description{
A dataset containing 704 cells and 24224 genes from mouse ESC
sce_mESC
}
\references{
Kolodziejczyk et al.
}
\keyword{datasets}
