% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segList}
\alias{segList}
\title{Stably expressed gene list}
\format{A list includes the stably expressed genes from human and mouse}
\usage{
segList
}
\description{
Stably expressed gene list
}
\keyword{datasets}
