% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segList_ensemblGeneID}
\alias{segList_ensemblGeneID}
\title{Stably expressed gene list in EnsemblGeneID}
\format{A list includes the stably expressed genes from human and mouse}
\usage{
segList_ensemblGeneID
}
\description{
Stably expressed gene list
}
\keyword{datasets}
