% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMatMult}
\alias{eigenMatMult}
\title{fast matrix multiplication}
\usage{
eigenMatMult(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}
}
\description{
fast matrix multiplication
}
\examples{
A = matrix(0, ncol = 1000, nrow = 1000)
system.time(A \%*\% A)
system.time(eigenMatMult(A, A))
}
