% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRUVIII.R
\name{fastRUVIII}
\alias{fastRUVIII}
\title{fastRUVIII}
\usage{
fastRUVIII(Y, M, ctl, k = NULL, eta = NULL, propEigen = 0.1,
  include.intercept = TRUE, average = FALSE, fullalpha = NULL,
  return.info = FALSE, inputcheck = TRUE)
}
\description{
perform fastRUVIII
}
\author{
Yingxin Lin
}
