% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvSimulate.R
\name{ruvSimulate}
\alias{ruvSimulate}
\title{Simulate a simple matrix}
\usage{
ruvSimulate(m = 100, n = 1e+05, nc = 1000, nRep = 50)
}
\arguments{
\item{m}{Number of observations}

\item{n}{Number of features}

\item{nc}{Number of negative controls}

\item{nRep}{Number of replicates}
}
\description{
Simulate a simple matrix
}
