% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvSimulate.R
\name{ruvSimulate}
\alias{ruvSimulate}
\title{Simulate a simple matrix to test the speed of scMerge}
\usage{
ruvSimulate(m = 100, n = 1e+05, nc = 1000, nRep = 50,
  lambda = 0.1, sce = FALSE)
}
\arguments{
\item{m}{Number of observations}

\item{n}{Number of features}

\item{nc}{Number of negative controls}

\item{nRep}{Number of replicates}

\item{lambda}{Rate parameter for random Poisson generation}

\item{sce}{If \code{TRUE}, returns a SingleCellExperiment object}
}
\description{
This function is designed to generate Poisson-random-variable data matrix
to test on the speed and performance of the internal algorithms of scMerge.
It is not intended to be used by end0-users.
}
