% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_cbind.R
\name{sce_cbind}
\alias{sce_cbind}
\title{Combind several \code{SingleCellExperiment} objects from different batches/experiments}
\usage{
sce_cbind(sce_list, method = NULL, cut_off_batch = 0.01,
  cut_off_overall = 0.01, exprs = c("counts", "logcounts"),
  colData_names = NULL, batch_names = NULL)
}
\arguments{
\item{sce_list}{A list contains the \code{SingleCellExperiment} Object from each batch}

\item{method}{A string indicates the method of combining the gene expression matrix,
either \code{union} or \code{intersect}}

\item{cut_off_batch}{A numeric vector indicating the cut-off for the proportion of a gene is expressed within each batch}

\item{cut_off_overall}{A numeric vector  indicating the cut-off for the proportion of a gene is expressed overall data}

\item{exprs}{A string vector indicating the exprssion matrices to be combined. The first assay named will be used to determine the proportion of zeores.}

\item{colData_names}{A string vector indicating the \code{colData} that are combined}

\item{batch_names}{A string vector indicating the batch names for the output sce object}
}
\value{
A combined \code{SingleCellExperiment} objects
}
\description{
Combind several \code{SingleCellExperiment} objects from different batches/experiments
}
\examples{
require(SingleCellExperiment)
data("sce_mESC", package = "scMerge.data")
batchNames<-unique(sce_mESC$batch)
sce_list<-list(sce_mESC[,sce_mESC$batch==batchNames[1]],
               sce_mESC[,sce_mESC$batch==batchNames[2]],
               sce_mESC[,sce_mESC$batch==batchNames[3]],
               sce_mESC[,sce_mESC$batch==batchNames[4]],
               sce_mESC[,sce_mESC$batch==batchNames[5]])
sce_combine<-sce_cbind(sce_list,batch_names=batchNames)
}
\author{
Yingxin Lin
}
