% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRUVg.R
\name{scRUVg}
\alias{scRUVg}
\title{RUVg function for single cell}
\usage{
scRUVg(Y, ctl, k, Z = 1, eta = NULL, include.intercept = TRUE,
  fullW = NULL, svdyc = NULL)
}
\description{
Modified based on RUV2 from package ruv and RUVg from package RUVseq function
}
\author{
Yingxin Lin
}
