% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRUVIII.R
\name{scRUVIII}
\alias{scRUVIII}
\title{scRUVIII}
\usage{
scRUVIII(Y = Y, M = M, ctl = ctl, fullalpha = NULL, k = k,
  return.info = TRUE, cell_type = NULL, batch = NULL,
  return_all = F, fast_svd = FALSE)
}
\description{
perform scRUVIII
}
\author{
Yingxin Lin
}
