% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segList}
\alias{segList}
\title{Stably expressed gene list in official gene symbols 
for both human and mouse}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(segList, package = 'scMerge')
}
\description{
A list includes the stably expressed genes 
for both human and mouse
}
\keyword{datasets}
